Fix for mms recieve on supersonic
mms recieve on supersonic is not working due to changes
in  content disposition parsing in PduParser.java.
This change fixes the if statement which previously ran
if the preference config_mms_content_disposition_support
was true, which it always was.
This change sets the preference in the supersonic
overlay config.xml to be false.

Change-Id: I2a9bb151c8c4f7308ce5dc36e174cc92eea3e834