supersonic: Light sensor adjustment to 10 levels, auto-brightness fixup
The kernel driver reports the following values for the light sensor:
3, 7, 12, 57, 114, 279, 366, 453, 540, 1023
This fix adjusts LightSensor.cpp by adding these values to the light sensor
index and getting all 10 values instead of the 4 we were getting before
(it appears other phone models will also need this adjustment).
The values are also used to set the auto brightness levels in config.xml,
and the button backlights are turned on for all of the levels.
This results in a functional automatic backlight once you check the
decrease light.  Once this fix makes it to other phones, we should
probably decrease brightness by default.

Change-Id: I596a48873e46ae8fef815afd38957f0f50809e28