supersonic: Match light sensor values to kernel driver values
The kernel driver uses the following values for the light sensor:
3, 7, 12, 57, 114, 279, 366, 453, 540, 1023
This fix adjusts LightSensor.cpp by adding these values to the light sensor
lux value index and getting 9 of the 10 values (max brightness 1023 seems to
be limited by the kernel driver). The values are also used to set the auto
brightness levels in config.xml, and the button backlights are turned on
for five of the levels. This results in a functional automatic backlight
once you check the decrease light option.

Update: changed the button backlight to hide at higher levels.

Change-Id: I5e3f6b907a917b7aeca478fa59317e43e30b3f43