Notification LED changes for EVO:  - Wimax LED is now used for battery info. This allows the tri-state android reporting (red for charging w/ low battery, orange for normal charging, green for full charge).  - liblights modified to allow routing to both speaker and wimax LEDs.  - CMParts configuration allows selection of both LEDs now.
Ideally we would allow the user to select from the several blink options available on the wimax LED (and some have requested solid/non-blinking LED support - see issue 2312), but I don't see any easy way of adding this without modifying CMParts itself. This may be done in a later update.

LED and the blink patterns can be selected outside of CMParts in apps that support it (such as Handcent's Power LED options)
 - Red/green/amber are routed to the wimax LED.
  - Any blinking of the red wimax LED will end up with the alternating green/red slow blink.
  - Blinking of the green wimax LED will be as follows: onMS <= 750 and offMS < onMS = two quick green flashes. onMS <= 750 and offMS >= onMS = one quick green flash. All else gives the slow green flash.
  - Blinking of amber/yellow is always a slow flash.
 - Mix blue in with the color to route to the speaker LED. Blue channel is ignored, and LED functionality is as it was before.

Change-Id: I2f9100e716a7821ad56abc781e08436df07fd9ad